feat(pure-json): stand up package with the JSON-faithfulness checker
Extract the value-level "unfaithful JSON" checker (findJsonUnfaithfulValues
+ JsonUnfaithfulValue + its private helpers) out of packages/llm/src/
schema-transport.ts into a new @commonfabric/pure-json package, and move its
tests along with it. The llm-specific assertJsonTransportSafe wrapper stays in
llm, re-importing the finder from pure-json; client.ts is untouched.
