+++ packages/pure-json/src/index.ts
@@ -1,0 +1,4 @@
+export {
+  findJsonUnfaithfulValues,
+  type JsonUnfaithfulValue,
+} from "./json-faithfulness.ts";
